requirements.txt
pytest == 8.4.1
randominfo~=2.0.2
pytz~=2025.2
pillow~=11.3.0
Flask~=3.1.1
requests~=2.32.4
SQLAlchemy~=2.0.42
Flask-JWT-Extended~=4.7.1
selenium~=4.34.2
psycopg2-binary~=2.9.10
allure-python-commons~=2.15.0